backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pymongo==4.8.0
motor==3.5.0
transformers==4.44.0
torch==2.4.0
Pillow==10.4.0
reportlab==4.2.0
python-multipart==0.0.9
exifread==3.0.0
networkx==3.3
python-dotenv==1.0.1